chore: moved filesystem util-functions to dedicated file
