+++ src/oxl_ansible_webui/aw/utils/util.py
@@ -3,9 +3,6 @@ import re as regex
 from platform import python_version
 from datetime import datetime, timedelta
 from time import time
-from os import open as open_file
-from os import remove as remove_file
-from pathlib import Path
 from functools import lru_cache, wraps
 from math import ceil
 from sys import maxunicode
@@ -75,58 +72,8 @@ def get_next_cron_execution_str(schedule: str, wait_sec: (int, float) = None) ->
     return next_exec_dt.strftime(SHORT_TIME_FORMAT)
 
 
-def _open_file_0600(path: (str, Path), flags):
-    return open_file(path, flags, 0o600)
-
-
-def write_file_0600(file: (str, Path), content: str):
-    file = Path(file)
-    if not file.parent.is_dir():
-        file.parent.mkdir(mode=0o750, parents=True, exist_ok=True)
-
-    mode = 'w'
-    if file.is_file():
-        mode = 'a'
-
-    with open(file, mode, encoding='utf-8', opener=_open_file_0600) as _file:
-        _file.write(content)
-
-
-def _open_file_0640(path: (str, Path), flags):
-    return open_file(path, flags, 0o640)
-
-
-def write_file_0640(file: (str, Path), content: str):
-    file = Path(file)
-    if not file.parent.is_dir():
-        file.parent.mkdir(mode=0o750, parents=True, exist_ok=True)
-
-    mode = 'w'
-    if file.is_file():
-        mode = 'a'
-
-    with open(file, mode, encoding='utf-8', opener=_open_file_0640) as _file:
-        _file.write(content)
-
-
 def get_random_str(l: int = 50) -> str:
     return ''.join(random_choice(ascii_letters + digits + punctuation) for _ in range(l))
-
-
-def overwrite_and_delete_file(file: (str, Path)):
-    if not isinstance(file, Path):
-        file = Path(file)
-
-    if not file.is_file():
-        return
-
-    for _ in range(3):
-        write_file_0600(
-            file=file,
-            content=get_random_str(),
-        )
-
-    remove_file(file)
 
 
 def get_ansible_versions() -> str:
